refactor(question): extract QueueService from QuestionService
Move the moderator/admin "Queue Details" rendering (getQueueSection,
getQueueDetails and their private mapping helpers) into a new QueueService via
facade + delegation. The gate-keeper/auditor role-status constants, shared with
QuestionService's role crons, are extracted to services/helpers/roleStatuses.ts;
the moderator cron now uses the resolveExpertMeta helper directly. QueueService
touches no allocation state (only 3 read repos). QuestionService: ~8,160 -> ~7,165 lines.
