+++ backend/src/modules/core/container.ts
@@ -19,6 +19,7 @@ import { PaeValidationService } from '../question/services/PaeValidationService.
 import { FeedbackService } from '../question/services/FeedbackService.js';
 import { QuestionAiService } from '../question/services/QuestionAiService.js';
 import { DuplicateService } from '../question/services/DuplicateService.js';
+import { QueueService } from '../question/services/QueueService.js';
 import { ContextController } from '../context/controllers/ContextController.js';
 import { ContextService } from '../context/services/ContextService.js';
 import { PerformanceController } from '../performance/controllers/PerformanceController.js';
@@ -65,6 +66,10 @@ export const coreContainerModule = new ContainerModule(options => {
   options
     .bind(GLOBAL_TYPES.DuplicateService)
     .to(DuplicateService)
+    .inSingletonScope();
+  options
+    .bind(GLOBAL_TYPES.QueueService)
+    .to(QueueService)
     .inSingletonScope();
   options.bind(CORE_TYPES.AnswerService).to(AnswerService).inSingletonScope();
   options.bind(CORE_TYPES.ContextService).to(ContextService).inSingletonScope();
